(refactor) Move commit chart code into its own file
