+++ client/index.html
@@ -5,6 +5,10 @@
       rect {
         fill: #2ae;
       }
+      svg {
+        font: 10px sans-serif;
+        font-size: 12px;
+      }
     </style>
     <title>GitAchieve</title>
   </head>
